pyproject.toml
[project]
name = "mleproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "plotly>=6.4.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
]
